requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
boto3==1.34.118
botocore==1.34.118
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-database-url==2.1.0
Django==5.0.4
django-modeltranslation==0.19.4
django-storages==1.14.3
django-taggit==5.0.1
forex-python==1.8
gunicorn==22.0.0
idna==3.7
jmespath==1.0.1
kombu==5.3.7
packaging==24.0
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycountry==24.6.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.6
requests==2.32.3
s3transfer==0.10.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
stripe==9.9.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
